==> competencia_03/Entrega Final/01_Clase_ternaria_competencia_3_SQL.ipynb <==
Colab para generar clase_ternaria y obtener el archivo competencia_02.csv.gz para subir al bucket de google cloud (extraido de la ayudita que dio el profe Alejandro en Zullip)

## IMPORTANTE: SUBIR EL ARCHIVO competencia_02_crudo.csv.gz al entorno del Google Colab

%%bash
pip install --upgrade duckdb
pip install --upgrade jupysql
pip install --upgrade duckdb-engine

"A diferencia de la competencia 1, ahora vamos a usar una base de datos en disco. Todo va a ser más lento, pero se va a poder trabajar sin problemas en máquinas de menor capacidad. La DB va a queda en el archivo dmeyf.db, lo probé en un colab básico y funcionó sin problemas."


import duckdb
import pandas as pd

%load_ext sql
%config SqlMagic.autopandas = True
%config SqlMagic.feedback = False
%config SqlMagic.displaycon = False
%sql duckdb:////content/dmeyf.db

"Cargamos el dataset, y como tenemos la base en el disco, no debemos cargarlo más, la próxima se puede usar directamente"


from google.colab import drive
drive.mount('/content/drive')

%%sql
create or replace table competencia_02 as
select
    *
from read_csv_auto('/content/drive/MyDrive/DM/DMEyF/competencia_02_crudo.csv.gz')


"El ayudin. Que hace el siguiente código:

Toma todos los valores de foto_mes y numero_de_cliente
Cruza todos con todos. De esta manera nos aseguramos que va haber un registro para todos los clientes en todos los periodos.
Luego se crean las variables soporte mes_0, que nos dice si el cliente esta en esa foto_mes, mes_1 si esta en la foto_mes siguiente y mes_2 si esta en las siguientes dos foto_mes
Debe completar su lógica en la variable clase_ternaria usando estas variables auxiliares
La tabla final se queda solamente con las variables y los registros necesarios."


#### Lógica que utilicé: Si mes_0 = 1, mes_1 = 0, mes_2 = 0  entonces BAJA +1 / Si mes_0=1, mes_1=1, mes_2 = 0, entonces BAJA+2 / Si mes_0=1, mes_1 = 0 , mes_2= 1 entonces CONTINUA / los demas casos son CONTINUA

%%sql
create or replace table targets as
with periodos as (
    select distinct foto_mes from competencia_02
), clientes as (
    select distinct numero_de_cliente from competencia_02
), todo as (
    select numero_de_cliente, foto_mes from clientes cross join periodos
), clase_ternaria as (
    select
        t.numero_de_cliente
        , t.foto_mes
        , if(c.numero_de_cliente is null, 0, 1) as mes_0
        , lead(mes_0, 1) over (partition by t.numero_de_cliente order by foto_mes) as mes_1
        , lead(mes_0, 2) over (partition by t.numero_de_cliente order by foto_mes) as mes_2
        , CASE
            WHEN mes_0 = 1 AND mes_1 = 0 AND mes_2 = 0 THEN 'BAJA+1'
            WHEN mes_0 = 1 AND mes_1 = 1 AND mes_2 = 0 THEN 'BAJA+2'
            WHEN mes_0 = 1 AND mes_1 = 0 AND mes_2 = 1 THEN 'CONTINUA'
            ELSE 'CONTINUA'
          END as clase_ternaria
    from todo t
    left join competencia_02 c using (numero_de_cliente, foto_mes)
) select
  foto_mes
  , numero_de_cliente
  , clase_ternaria
from clase_ternaria where mes_0 = 1


"Luego agregamos la variable clase_ternaria a nuestra tabla y cargamos los datos que previamente calculamos."


%%sql
alter table competencia_02 add column clase_ternaria VARCHAR(10)


%%sql
update competencia_02
set clase_ternaria = targets.clase_ternaria
from targets
where competencia_02.numero_de_cliente = targets.numero_de_cliente and competencia_02.foto_mes = targets.foto_mes


"Y finalmente guardamos el dataset para usarlo en la competencia. Esto les va a tardar en guardar."


%%sql
copy competencia_02 to '/content/competencia_02.csv.gz' (FORMAT CSV, HEADER)

#### Importante: Hay que descargarse el archivo "competencia_02.csv.gz" generado y subirlo al bucket de google cloud

##### Para chequear que esté bien creada clase_ternaria

%%sql
create or replace table competencia_02 as
select
    *
from read_csv_auto('/content/competencia_02.csv.gz')

%%sql
SELECT * FROM competencia_02 WHERE clase_ternaria IS NOT NULL LIMIT 200
